+++ src/wfg/ltv_econ/configs/LaborConfigLoader.java
@@ -1,7 +1,12 @@
 package wfg.ltv_econ.configs;
 import static wfg.ltv_econ.constants.economyValues.MONTH;
+
+import java.util.HashMap;
+import java.util.Map;
+
 import static wfg.ltv_econ.constants.Mods.*;
 
+import org.json.JSONArray;
 import org.json.JSONObject;
 
 import com.fs.starfarer.api.Global;
@@ -12,6 +17,8 @@ public class LaborConfigLoader {
     private static final String CONFIG_PATH = "./data/config/ltvEcon/labor_config.json";
 
     private static JSONObject config;
+
+    public static final String AVERAGE_OCC_TAG = "average"; 
 
     private static final void load() {
         try {
@@ -37,15 +44,16 @@ public class LaborConfigLoader {
         LaborConfig.avg_wage = LaborConfig.LPV_month / LaborConfig.RoSV;
         LaborConfig.defaultWorkerCapPerOutput = (float) root.getDouble("defaultWorkerCapPerOutput");
         LaborConfig.dynamicWorkerCapPerOutput = (float) root.getDouble("dynamicWorkerCapPerOutput");
-        LaborConfig.RoVC_average = (float) root.getDouble("RoVC_average");
-        LaborConfig.RoVC_industry = (float) root.getDouble("RoVC_industry");
-        LaborConfig.RoVC_consumer = (float) root.getDouble("RoVC_consumer");
-        LaborConfig.RoVC_manufacture = (float) root.getDouble("RoVC_manufacture");
-        LaborConfig.RoVC_service = (float) root.getDouble("RoVC_service");
-        LaborConfig.RoVC_agriculture = (float) root.getDouble("RoVC_agriculture");
-        LaborConfig.RoVC_mechanized = (float) root.getDouble("RoVC_mechanized");
-        LaborConfig.RoVC_manual = (float) root.getDouble("RoVC_manual");
-        LaborConfig.RoVC_space = (float) root.getDouble("RoVC_space");
+
+        final JSONArray RoVCList = root.getJSONArray("RoVC_list");
+        for (int i = 0; i < RoVCList.length(); i++) {
+            final JSONObject RoVCObj = RoVCList.getJSONObject(i);
+            final float value = (float) RoVCObj.getDouble("value");
+            final String type = RoVCObj.getString("type");
+            LaborConfig.RoVC_map.put(type, value);
+
+            if (type.equals(AVERAGE_OCC_TAG)) LaborConfig.RoVC_average = value;
+        }
 
         if (Global.getSettings().getModManager().isModEnabled(LUNA_LIB)) {
             LaborConfig.RoSV = LunaSettings.getInt(LTV_ECON, "labor_RoSV");
@@ -53,21 +61,23 @@ public class LaborConfigLoader {
             LaborConfig.LPV_month = LunaSettings.getInt(LTV_ECON, "labor_lpvMonth");
             LaborConfig.LPV_day = LaborConfig.LPV_month / (float) MONTH;
             LaborConfig.avg_wage = LaborConfig.LPV_month / LaborConfig.RoSV;
-            LaborConfig.RoVC_average = LunaSettings.getDouble(LTV_ECON, "labor_RoVC_average").floatValue();
-            LaborConfig.RoVC_industry = LunaSettings.getDouble(LTV_ECON, "labor_RoVC_industry").floatValue();
-            LaborConfig.RoVC_consumer = LunaSettings.getDouble(LTV_ECON, "labor_RoVC_consumer").floatValue();
-            LaborConfig.RoVC_manufacture = LunaSettings.getDouble(LTV_ECON, "labor_RoVC_manufacture").floatValue();
-            LaborConfig.RoVC_service = LunaSettings.getDouble(LTV_ECON, "labor_RoVC_service").floatValue();
-            LaborConfig.RoVC_agriculture = LunaSettings.getDouble(LTV_ECON, "labor_RoVC_agriculture").floatValue();
-            LaborConfig.RoVC_mechanized = LunaSettings.getDouble(LTV_ECON, "labor_RoVC_mechanized").floatValue();
-            LaborConfig.RoVC_manual = LunaSettings.getDouble(LTV_ECON, "labor_RoVC_manual").floatValue();
-            LaborConfig.RoVC_space = LunaSettings.getDouble(LTV_ECON, "labor_RoVC_space").floatValue();
+
+            final float avgValue = LunaSettings.getDouble(LTV_ECON, "labor_RoVC_average").floatValue();
+            LaborConfig.RoVC_average = avgValue;
+            LaborConfig.RoVC_map.put(AVERAGE_OCC_TAG, avgValue);
+            LaborConfig.RoVC_map.put("industry", LunaSettings.getDouble(LTV_ECON, "labor_RoVC_industry").floatValue());
+            LaborConfig.RoVC_map.put("consumer", LunaSettings.getDouble(LTV_ECON, "labor_RoVC_consumer").floatValue());
+            LaborConfig.RoVC_map.put("manufacture", LunaSettings.getDouble(LTV_ECON, "labor_RoVC_manufacture").floatValue());
+            LaborConfig.RoVC_map.put("service", LunaSettings.getDouble(LTV_ECON, "labor_RoVC_service").floatValue());
+            LaborConfig.RoVC_map.put("agriculture", LunaSettings.getDouble(LTV_ECON, "labor_RoVC_agriculture").floatValue());
+            LaborConfig.RoVC_map.put("mechanized", LunaSettings.getDouble(LTV_ECON, "labor_RoVC_mechanized").floatValue());
+            LaborConfig.RoVC_map.put("manual", LunaSettings.getDouble(LTV_ECON, "labor_RoVC_manual").floatValue());
+            LaborConfig.RoVC_map.put("space", LunaSettings.getDouble(LTV_ECON, "labor_RoVC_space").floatValue());
         }
 
         } catch (Exception e) {
         throw new RuntimeException(
-            "Failed to load labor configuration from " + CONFIG_PATH + ": "
-            + e.getMessage(), e
+            "Failed to load labor configuration from " + CONFIG_PATH, e
         );
         }
     }
@@ -84,77 +94,18 @@ public class LaborConfigLoader {
         public static float dynamicWorkerCapPerOutput;
 
         public static float RoVC_average;
-        public static float RoVC_industry;
-        public static float RoVC_consumer;
-        public static float RoVC_manufacture;
-        public static float RoVC_service;
-        public static float RoVC_agriculture;
-        public static float RoVC_mechanized;
-        public static float RoVC_manual;
-        public static float RoVC_space;
+        public static final Map<String, Float> RoVC_map = new HashMap<>();
 
         static {
             LaborConfigLoader.loadConfig();
         }
 
-        public static final float getRoVC(OCCTag tag) {
-            switch (tag) {
-            case INDUSTRY:
-                return RoVC_industry;
-            case CONSUMER:
-                return RoVC_consumer;
-            case MANUFACTURE:
-                return RoVC_manufacture;
-            case SERVICE:
-                return RoVC_service;
-            case AGRICULTURE:
-                return RoVC_agriculture;
-            case MANUAL:
-                return RoVC_manual;
-            case MECHANIZED:
-                return RoVC_mechanized;
-            case SPACE:
-                return RoVC_space;
-            case AVERAGE:
-            default:
-                return RoVC_average;
-            }
+        public static final float getRoVC(final String tag) {
+            return RoVC_map.getOrDefault(tag, RoVC_average);
         }
 
-        public static final float getRoCC(OCCTag tag) {
-            switch (tag) {
-            case INDUSTRY:
-                return 1f - RoVC_industry;
-            case CONSUMER:
-                return 1f - RoVC_consumer;
-            case MANUFACTURE:
-                return 1f - RoVC_manufacture;
-            case SERVICE:
-                return 1f - RoVC_service;
-            case AGRICULTURE:
-                return 1f - RoVC_agriculture;
-            case MECHANIZED:
-                return 1f - RoVC_mechanized;
-            case MANUAL:
-                return 1f - RoVC_manual;
-            case SPACE:
-                return 1f - RoVC_space;
-            case AVERAGE:
-            default:
-                return 1f - RoVC_average;
-            }
+        public static final float getRoCC(final String tag) {
+            return 1f - RoVC_map.getOrDefault(tag, RoVC_average);
         }
-    }
-
-    public static enum OCCTag {
-        AVERAGE,
-        INDUSTRY,
-        CONSUMER,
-        MANUFACTURE,
-        SERVICE,
-        AGRICULTURE,
-        MECHANIZED,
-        MANUAL,
-        SPACE
     }
 }